Rename ProgressBar file
