+++ src/components/Application/ProgressArea.tsx
