+++ src/components/Application/index.ts
@@ -2,7 +2,7 @@ import ApplicationContainer from './ApplicationContainer'
 import ApplicationSections from './ApplicationSections'
 import ApplicationSelectType from './ApplicationSelectType'
 import Navigation from './Navigation'
-import ProgressArea from './ProgressBar'
+import ProgressArea from './ProgressArea'
 
 export {
   ApplicationContainer,
